+++ projects/ng-translatable/src/lib/translatable-option.model.ts
@@ -1,0 +1,10 @@
+/* 3rd party libraries */
+
+/* locally accessible feature module code, always use relative path */
+
+export interface TranslatableOption {
+
+  readonly value: string;
+  readonly text: string;
+  selected: boolean;
+}