+++ projects/ng-translatable/src/projects.ts
@@ -1,0 +1,10 @@
+/*
+ * Public API Surface of ng-translatable
+ */
+
+export * from './lib/translatable-text-list.model';
+export * from './lib/translatable-option-list.model';
+export * from './lib/translatable-language-list.model';
+export * from './lib/translatable-multiple-choice.model';
+
+export * from './lib/translatable-option.model';